+++ studio/cloud_worker_util.py
@@ -1,6 +1,6 @@
 import os
 
-from .util import rand_string
+from util.util import rand_string
 """
 Utility functions for anything shared in common by ec2cloud_worker and
 gcloud_worker
