+++ studio/encrypted_payload_builder.py
@@ -8,11 +8,11 @@ import base64
 import json
 from sshpubkeys import SSHKey
 
-from .ed25519_key_util import Ed25519KeyUtil
+from studio.ed25519_key_util import Ed25519KeyUtil
 from payload_builders.payload_builder import PayloadBuilder
 from studio import logs
-from .unencrypted_payload_builder import UnencryptedPayloadBuilder
-from .util import check_for_kb_interrupt
+from studio.unencrypted_payload_builder import UnencryptedPayloadBuilder
+from util.util import check_for_kb_interrupt
 
 class EncryptedPayloadBuilder(PayloadBuilder):
     """
