+++ studio/gpu_util.py
@@ -2,7 +2,7 @@ import os
 import subprocess
 import xml.etree.ElementTree as ET
 
-from .util import sixdecode
+from util.util import sixdecode
 
 
 def memstr2int(string):
